refactor: Migrate legacy components to feature-based structure
This commit migrates all components from the deprecated src/components/ directory
to the appropriate feature-based structure as outlined in CLAUDE.md guidelines.

**Changes made:**

**Landing Page Components** (moved to `features/landing/components/`):
- Contact.tsx
- Features.tsx
- HeroSection.tsx
- HowBanner.tsx
- LandingFooter.tsx
- Reviews.tsx
- WhyBanner.tsx

**Campaign Forms** (moved to `features/campaign/components/`):
- AudienceDetails.tsx
- AudienceForm.tsx
- CampaignBriefForm.tsx (+ test file)
- PurposeForm.tsx (+ test file)
- TimingForm.tsx

**Generic UI Components** (moved to `shared/components/ui/`):
- CardsSection.tsx
- ConnectToInstagram.tsx
- ExportCompleteDialog.tsx
- ExportLoading.tsx
- ImageStack.tsx
- InstagramIcons.tsx
- PhysicalLocations.tsx
- PostHogPageviewTracker.tsx

**Import Path Updates:**
- Updated all import statements across the codebase to reference new component locations
- Fixed test file imports for moved components
- Verified all imports resolve correctly

**Quality Assurance:**
- All tests pass successfully
- No linting errors
- Code formatted with Prettier
- Follows feature-based architecture principles from CLAUDE.md

This refactoring improves code organization, maintainability, and scalability by:
1. Grouping related components with their features
2. Creating clear boundaries between feature-specific and shared components
3. Making it easier to locate and maintain components
4. Following established architectural patterns
