+++ src/features/campaign/components/PostPreview.tsx
@@ -1,7 +1,7 @@
 import React, { useContext, useEffect, useRef, useState } from 'react';
 import PaginationImage from './PaginationImage';
 import { BsThreeDots } from 'react-icons/bs';
-import InstagramIcons from '../../../components/InstagramIcons';
+import InstagramIcons from '../../../shared/components/ui/InstagramIcons';
 import Options from '../../../shared/components/ui/Options';
 import { BrandContext } from '../../brand/context/BrandContext';
 import ModifyPromptForm from './ModifyPromptForm';
