+++ src/shared/components/ui/CardsSection.tsx
