+++ src/shared/components/ui/ImageStack.tsx
